move store folder to the app folder
